Makes autovalue bean constructor always visible
Changes the visibility of the constructor for the generated bean to be
unconditionally package private.

Now it is possible to have static construction methods and a builder
at the same time e.g. when creating a bean incrementally in logic
while deserializing it wholesale from e.g. JSON or Protobuf.