Support for TYPE_USE @Nullable.
Limitations: The following cases currently don't work with javac (i.e.
the annotation processors don't see the @Nullable so a null check is
generated in the constructor):

public abstract int @Nullable [] getInts()
This works with eclipse JDT and the null check is ommited in the
generated code, but note that the @Nullable is in the wrong position in
the generated code (i.e. @Nullable int[] instead of int @Nullable[])),
which might lead to compiler warnings

For generic type T: 
@Nullable
public abstract T getT()
(works in JDT)

Also @Nullable won't work for types which can't be imported because of
naming conflicts, because the @Nullable is generated at the wrong place,
i.e. "@Nullable some.package.Map" instead of "some.package.@Nullable
Map" which won't compile at all for TYPE_USE @Nullable.