Generalize support for property builders.
Now, in addition to being able to say immutableListBuilder() for a property of type ImmutableList<T>, you can say fooBuilder() for a property of an arbitrary type that has a builder following certain conventions. In particular, you can do this if the type of foo() is itself an @AutoValue class with a builder.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=140584023