In generated Builder classes, make the second (toBuilder) constructor private.
This constructor was never documented, and its presence means that we can't know whether @AutoValue instances are converted back into builders or not. That in turn means that, currently, if we have `Foo foo()` in the @AutoValue class and `FooBuilder fooBuilder()` in the @AutoValue.Builder class, then `Foo` must have a `FooBuilder toBuilder()` method, in case a `Foo` ever needs to be converted back into a `FooBuilder`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=140958480