Explicitly narrow long to int before adding to hash.
The compound assignment to the hash value already performed the integer narrowing but did so implicitly. Error-prone was recently updated to detect this case and thus marked this as problematic. Rasterizing the implicit cast as explicitly makes the intent clear and resolves the warnings.

This behavior is already present in AutoAnnotation but was missing from AutoValue.