Suppress some sources of warnings in AutoOneOfCompilationTest, to avoid a test failure that only shows up on Travis.
I'm still trying to understand why we get these failures in the first place. My theory is that, in this environment only, the test is finding com/google/auto/value/AutoOneOf.java when it sees `import com.google.auto.value.AutoOneOf`, so that file is "implicitly compiled". I'm not sure if that might be considered a bug in compile-testing, or in the Travis configuration.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183085105