Copy type annotations to the parameter of equals(Object).
If equals(Object) is being implemented because there is abstract redeclaration of it in the AutoValue class or an ancestor, and if that redeclaration has a type annotation such as @Nullable on its parameter, then the generated implementation will also have that annotation. The same applies to AutoOneOf.

We could go further and also copy annotations on the return types of equals/hashCode/toString, but for now there's no perceived need for that.

Fixes https://github.com/google/auto/issues/579.

RELNOTES=In `@AutoValue`, copy `@Nullable` from `equals(@Nullable Object)` to its implementation.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=188044994