Make auto-value an incremental processor (in Gradle)
Uses net.ltgt incap-processor (an auto-service-like processor)
to generate the META-INF file used by Gradle.

The AutoValue processor is marked as "dynamic" and computes
whether it really is incremental by asking extensions. The
processor is "isolating" by default, in the absence of any
extension; and extensions default to "unknown", which turn
the processor to non-incremental (the MemoizeExtension,
which is always present, is marked as "isolating").

AutoAnnotation, AutoOneOf, and AutoValueBuilder processors
are marked as "isolating", as they don't support extensions.

Also, removes OmitIdentifiers option from AutoOneOf processor
as it doesn't actually makes anything with it.