Make it a compilation error for an @AutoOneOf property to be @Nullable.
The generated code already rejects null values, whether or not @Nullable is present, so this just lets users know earlier that what they are trying is not going to work.

RELNOTES=It is a compilation error for an @AutoOneOf property to be @Nullable since null values are not in fact accepted.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=229442655