Further improve the logic for checking setter types.
In the case where the setter parameter is supposed to be the same as the getter type, we use appropriate logic to take inheritance into account. We also verify that the type of a parameter that we are thinking of feeding to ImmutableSet.copyOf (for example) will actually lead to an ImmutableSet of the right type.

RELNOTES=Improved type checking for builder setter parameters.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=238503171