Improve Maven instructions for AutoValue.
It is not true that a dependency on the auto-value artifact transitively implies a dependency on auto-value-annotations, so it is not correct to suggest that users can include only the former.

Make it clearer that <plugins> is inside the <build> section.

Use ${auto-value.version} consistently, rather than sometimes that and sometimes ${version}.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=283224151