If a getter is `getOAuth()`, allow the setter to be `oAuth(T)`.
Normally we follow the JavaBeans specification, which has special rules for properties that begin with more than one capital, like the `OAuth` implied by `getOAuth()` here. That means that the unprefixed setter should be spelled `OAuth(T)`. That's obviously obscure and unsettling, so we add logic to allow the setter to be either `OAuth(T)` or `oAuth(T)` in this case. We already have logic for the opposite case, where the getter is `oAuth()` and the setter is `setOAuth(T)`.

Fixes https://github.com/google/auto/issues/565.

RELNOTES=Avoid surprising behaviour when getters are prefixed and setters are not, and a property name begins with two capital letters.
PiperOrigin-RevId: 351188819