Add a simple ad-hoc indentation engine to the reformatting pass.
The background here is that AutoValue and its cousins produce code using Velocity templates, and then run a reformatter over the result to fix up various whitespace problems. That works fine except that it's very difficult to manage indentation in templates. For example, given something like this:

```
  ${class}(
  #foreach ($p in $properties)
      $p.type $p.name #if ($foreach.hasNext) , #end

  #end
  ) {
```

Velocity deletes newlines after directives like `#foreach`, so we will get the spaces before `#foreach` at the start of the first parameter line and we'll get the spaces before `#end` at the start of the line beginning with `)`. This particular case can be fixed by adding a newline after the `#foreach` and after the `#end`. But it's very tricky to manage all these extra blank lines.

What this change does is to replace whatever initial space comes out of the template with the correct indentation that we compute. That should allow us to make the templates a lot more readable by deleting the weird blank lines. The example above should work correctly without needing them.

RELNOTES=n/a
PiperOrigin-RevId: 363340310