Reduce the number of warnings from `CompileWithEclipseTest`.
There were 133 warnings, which made it hard to find any errors or unexpected warnings that might crop up. With this change there are 11 warnings, which we might later be able to reduce further.

Also fix potentially-raw types in an unusual case, `equals` for generic `@AutoValue` classes with primitive array fields.

RELNOTES=n/a
PiperOrigin-RevId: 364357181