Initial prototype of `@AutoBuilder`.
This is not ready for use yet.

For now, it only supports constructor invocation, not static methods; it does not handle overloads correctly (it requires there to be a maximal constructor and uses that); it has problems with case when getters have names like `getAString`. There are other problems.

RELNOTES=n/a
PiperOrigin-RevId: 366603722