Move knowledge of getter methods from `BuilderMethodClassifier` to a new subclass.
This prepares the sharing of `BuilderMethodClassifier` between AutoValue and AutoBuilder. The logic for handling builders is substantially the same between the two. The main difference is that AutoValue properties come from abstract getter methods in the `@AutoValue` class, while AutoBuilder properties come from the parameters of the constructor or method that will be called. Before this change we had `ExecutableElement` instances that were the getter methods. Now we have `Element` instances that are refined as `ExecutableElement` in the new subclass `BuilderMethodClassifierForAutoValue`. In a later step we will introduce `BuilderMethodClassifierForAutoBuilder` where each `Element` will each be a `VariableElement` representing the property parameter.

Earlier versions of this change ran into confusion between "original" and "rewritten" property types, so `BuilderMethodClassifier` now makes that distinction more explicit.

RELNOTES=n/a
PiperOrigin-RevId: 366888706