Have AutoBuilder determine property names and types from the parameters of the `callMethod`.
Previously, for convenience, we reused the logic from AutoValue of examining abstract getter methods in the built type. But AutoBuilder is not specified to do that. Instead, every parameter of the target constructor or `callMethod` method defines the name and type of a property, and the builder must specify a way to set that property.

RELNOTES=n/a
PiperOrigin-RevId: 367056519