Add overload-resolution logic to AutoBuilder.
There can be more than one visible constructor (or static method) in the `ofClass` class. This change adds some ad-hoc logic to pick the right one by matching the names of its parameters against the methods in the `@AutoBuilder` interface.

There is some duplication of knowledge between this new logic and the existing `BuilderMethodClassifier` logic. Another approach would be to try `BuilderMethodClassifier` for _each_ visible constructor and retain whichever one it works for. The concern there, though, is that if the user code is wrong it will be hard to produce succinct error messages explaining why. With the approach here, we can show for each constructor which parameters did not match (in a later change).

Also, the problem where we didn't handle a property called aString very well has been fixed, so adjust tests accordingly.

RELNOTES=n/a
PiperOrigin-RevId: 367237991