`MoreTypes.isTypeOf` returns false for `ErrorType` rather than throwing.
Throwing `IllegalArgumentException` can mask the real source of a problem, or prevent code that would have handled the `ErrorType` from running. The `ErrorType` typically (maybe always) corresponds to an identifier in the source code, so it is in fact a class, just one that doesn't exist.

RELNOTES=MoreTypes.isTypeOf returns false for ErrorType rather than throwing.
PiperOrigin-RevId: 369017594