Fix a problem where `@SerializableAutoValue` could generate incorrect code for complex types.
The problem is that the generated code can include nested lambdas. If an inner lambda uses the same identifier for a parameter as an outer one, that's a kind of hiding that the Java compiler doesn't allow. So we add a method to `SerializerFactory` that generates unique identifiers.

RELNOTES=Fixed a problem where `@SerializableAutoValue` could generate incorrect code for complex types.
PiperOrigin-RevId: 370078937