Clarify that `Optional` values can be set from `@Nullable` parameters.
This was not all that clear from the documentation. If you have `Optional<String> getFoo()` property then you can set it with `setFoo(String)` _or_ `setFoo(@Nullable String)`. In the first case, you'll get an exception if the parameter is null, but in the second it will work, and `foo` will be `Optional.empty()`.

Also add a test that checks that you do indeed get an exception without `@Nullable`.

RELNOTES=n/a
PiperOrigin-RevId: 375607596