Replace server Guava API usage with Android compatible alternatives.
The issue occurs when the Android flavor of Guava somehow finds its way onto the processor class path and gets classloaded instead of the server Guava. The ideal solution would be to just keep the Android flavor of Guava off of the processor class path. However, it's been something that we've allowed for a while now, e.g. in Dagger, and would be difficult to start enforcing.

This CL adds a new API for auto-common. A follow-up CL will update auto-value and auto-service to depend on the new auto-common APIs once auto-common has a release with those changes.

RELNOTES=Added new public APIs for guava collector usage (toImmutable*) that are Android-compatible.
PiperOrigin-RevId: 376892205