Upgrade dependencies to auto-common 1.0.1 and replace Guava toImmutable* usages with auto-common alternatives.
This fixes an issue that occurs when the Android flavor of Guava somehow finds its way onto the processor class path and gets classloaded instead of the server Guava. The ideal solution would be to just keep the Android flavor of Guava off of the processor class path. However, it's been something that we've allowed for a while now, e.g. in Dagger, and would be difficult to start enforcing.

RELNOTES=N/A
PiperOrigin-RevId: 376926489