Add String and TypeElement versions of auto-common APIs that currently take in Class.
These changes continue the effort to move away from referencing classes from the user API directly in processors, which can cause issues especially with Android dependencies.

Note: I decided not to do anything with isTypeOf yet because I'm not sure of a great way to support that.
RELNOTES=Add String and TypeElement versions of auto-common APIs that currently take in Class.
PiperOrigin-RevId: 377287397