Add support for printing `AnnotationValue`s and `AnnotationMirror`s
`AnnotationValue#toString` is specified to return a string "suitable for
representing this value in the source code of an annotation", but recent
versions of javac have started using simple names for types in annotations
and class literals instead of qualified names, which breaks annotation
processors that expect the source representation to be self-contained
(see https://bugs.openjdk.java.net/browse/JDK-8164819).

This change adds `AnnotationMirrors.toString` and `AnnotationValues.toString`, which format `AnnotationValue`s and `AnnotationMirror`s using qualified names
of types to ensure the types can be resolved in generated code.

RELNOTES=Add `AnnotationMirrors.toString` and `AnnotationValues.toString`
PiperOrigin-RevId: 385196293