Upgrade Gradle Test Kit to 7.2
It isn't exactly necessary to upgrade Gradle Test Kit to the latest
version as Gradle Test Kit as with normal Java dependencies. The Gradle
Test Kit JARs contains plumbing used to execute a Gradle build within as
part of testing scenarios. The Gradle distribution usually generate the
JARs at runtime. As the classes are released with each new Gradle version,
they may not change between versions. The `dev.gradleplugins` project
rerelease the Gradle Test Kit JARs to be used outside of a Gradle build
and with a different Gradle version. Unfortunately, we keep the Gradle
distribution versioning which leads to confusion about the "latest"
version. Generally, using the latest version is safe but keep in mind the
Gradle Test Kit JARs depend on the Gradle API, Groovy and Kotlin. If your
project has a strong opinion on those dependencies, you will need to
resolve the version conflict. Note that Groovy and Kotlin dependencies
are mostly for the DSL part of Gradle, which is unlikely to leak in Test
Kit but could via the implementation of the classes.