Refine the check for overriding interface methods.
One interface method overrides another only if the overrider is in a subinterface of the overridden. This is the first bullet in [JLS §9.4.1.1](https://docs.oracle.com/javase/specs/jls/se17/html/jls-9.html#jls-9.4.1.1).

Remove a test that was checking for the old behaviour. It was intended to fix a bug in Dagger, but Dagger now works fine without this fix, and has its own test for the relevant situation.

RELNOTES=In `MoreElements.overrides` and `.getLocalAndInheritedMethods`, an interface method is no longer considered to override another interface method unless the first interface inherits from the second. This means that `.getLocalAndInheritedMethods` may now return two methods with the same signature where before it only returned one.
PiperOrigin-RevId: 403430316