Handle missing type when copying annotations.
If you have `@CopyAnnotations` and `@Foo(Bar.class)`, and if `Bar` is undefined, we want to defer processing to give some other annotation processor a chance to define `Bar`. This turns out to be quite difficult, because javac essentially makes it look as if you have written `@Foo("<error>")` in this case. So we check to see if an annotation element that _should_ be a class is something else (a string in this case).

RELNOTES=We now handle better the case where an annotation being copied references a missing class.
PiperOrigin-RevId: 404307632