Make it easier to make a step-builder for AutoValue.
We no longer require each setter method in an `@AutoValue.Builder` class `Builder` to return `Builder`. A setter can return some supertype, in particular a step interface. That removes the need to override each inherited setter method just to change its return type.

Also add documentation showing how this works. Previously we just linked to an example in a GitHub issue.

RELNOTES=Making a step-builder for AutoValue is now easier because the inherited setters don't need to be restated in the `Builder` class so that they return `Builder`.
PiperOrigin-RevId: 407161124