Remove usage of Guava 30 API (Comparators.min).
This CL removes the only Guava 30 usage in auto-common, Comparators.min.

Motivation for change:
Currently when writing annotation processors in Bazel, the Guava version is pinned to the version used by JavaBuilder. For the latest Bazel release, 4.2.1, this version is pinned to Guava 29, which means we can't use the latest auto-common in our annotation processors due to the usage of Comparators.min.

Fwiw, Bazel 5.0 upgrades to Guava 30, but that version is still in the pre-release phase.

RELNOTES=Remove usage of Guava 30 API (Comparators.min).
PiperOrigin-RevId: 409461735