Validate that `@AutoValue` (etc) classes have appropriate constructors.
The class must have a non-private no-arg constructor so it can be subclassed in the generated code. Previously, if it didn't, we would generate code anyway and rely on the somewhat obscure compiler error message to signal the problem to the user. Now we check for the situation explicitly and produce a specific error message.

Fixes https://github.com/google/auto/issues/1209.

RELNOTES=n/a
PiperOrigin-RevId: 414779231