Behave better when property builders collide with JavaBeans eccentricity.
JavaBeans specifies that a method called getHTMLPage defines a property called HTMLPage, rather than hTMLPage as its regular decapitalization would suggest.  This is almost never what you want in practice but it's probably too late for us to stop following that rule. We've relaxed the matching in a number of places, so for example if your getter is getOAuth() then your setter can be oAuth(x).  Here we're extending that courtesy to property builders. If your AutoValue has a method `ImmutableList<String> getOAuths()` then its builder can continue to have a property builder `ImmutableList.Builder<String> OAuthsBuilder()`, but with this change that method can also be `ImmutableList.Builder<String> oAuthsBuilder()`.

(This change is probably too minor to go in the release notes.)

RELNOTES=n/a
PiperOrigin-RevId: 421830872