Make Serializable and Memoized extensions work together.
The Serializable extension has to call the constructor of the AutoValue class as part of its implementation of `readResolve`. The class whose constructor is called must be the *final* one in the chain of subclasses generated by extensions. That is not necessarily the subclass generated by the Serializable extension.

RELNOTES=Fixed an issue when Serializable and Memoized extensions are used together.
PiperOrigin-RevId: 430458407