Add a class that generates bytecode to call a constructor.
This will be used in a later step to call synthetic constructors. Those constructors can't be called from Java source code.

Introducing the ASM dependency caused problems with the modified Maven Shade Plugin we were using. The purpose of that plugin was to allow us to prefix `$` to the simple name of shaded classes, so they wouldn't be proposed in IDE completions. But we no longer recommend that people put the AutoValue processor jar in their compilation classpath, so we can switch back to the regular Maven Shade Plugin.

RELNOTES=n/a
PiperOrigin-RevId: 435426768