(Common) visiting -> visited, other minor improvements
I found the use of the word *visiting* confusing and not capturing what the code is doing. I changed it to *visited* instead, and other places where this word was relevant was also changed accordingly.

Moreover, to improve readability some nested if statements were flattened and some comments improved.

The flattening:
*** FROM ***
if( cond ){
    stmt(s)
    return expr
}
return false

*** TO ***
if( !cond )
    return false
stmt(s)
return expr