For `@AutoBuilder` classes, generate a "copy constructor" if appropriate.
For example, given this:

```
record Foo(String bar) {
  @AutoBuilder
  abstract class Builder {
    Builder setBar(String bar);
    Foo build();
  }
}
```

The built type is `Foo` (it is the return type of the `build()` method), and the record does have an implicit method `Foo.bar()`, so this qualifies. In addition to the `AutoBuilder_Foo_Builder()` constructor, we will declare `AutoBuilder_Foo_Builder(Foo source)`, which copies initial values for the builder from the given `source`.

RELNOTES=The generated AutoBuilder class now has a "copy constructor" if values for the builder properties can be obtained from the built type.
PiperOrigin-RevId: 448049949