Correctly handle `toBuilder()` when some but not all properties have defaults.
We have a bitmask of properties that have been set. When there are no Kotlin default properties, this bitmask covers only primitive properties. (For any other property, `null` means it hasn't been set.) When there *are* Kotlin default properties, the bitmask covers *all* properties, and it will be used to tell the Kotlin runtime which ones have been set (so the others can be defaulted). In the case of `toBuilder()`, *every* property is set, so we should set this bitmask to all-ones.

RELNOTES=n/a
PiperOrigin-RevId: 448255765