Unit-test for strange open-jdk behaviour
In the following example,

class ClassA{
	@Annotation
	void overloadedMethod(SomeGeneratedClass c) {}
	@Annotation
	void overloadedMethod(int c){}
}

at least open-jdk does not report the second method if `SomeGeneratedClass` references a `TypeKind#ERROR` when `RoundEnvironment#getElementsAnnotatedWith` is called, or even when `TypeElement#getEnclosedElements()` is called. Therefore, our implementation should be vigilant that the second method is captured for processing at some point.

Note that for a method to get "hidden" like this, it should reside after the `ERROR` referencing method, and it should not have any distinguishing characteristic like different name, different number of parameter, or a clear parameter type mismatch.